# Conflicts:
#	app/src/main/java/com/bink/wallet/scenes/add_auth/AddAuthFragment.kt
#	app/src/main/java/com/bink/wallet/scenes/loyalty_details/LoyaltyCardDetailsFragment.kt
#	app/src/main/res/layout/add_auth_fragment.xml
#	app/src/main/res/layout/card_item.xml
#	app/src/main/res/layout/fragment_loyalty_card_details.xml
#	app/src/main/res/values/dimens.xml

+++ app/src/main/java/com/bink/wallet/model/response/payment_card/Consent.kt
@@ -8,7 +8,7 @@ import kotlinx.android.parcel.Parcelize
 @JsonClass(generateAdapter = true)
 data class Consent(
     val type: Int,
-    val latitude: Double,
-    val longitude: Double,
+    val latitude: Float,
+    val longitude: Float,
     val timestamp: Long
 ) : Parcelable
